refactor(axvm): consolidate architecture-specific code
