+++ virtualization/axvm/src/arch/aarch64/boot/fdt/loongarch64/guest_firmware_dtb.rs
