+++ virtualization/axvm/src/arch/aarch64/boot/fdt/loongarch64/property.rs
