+++ virtualization/axvm/src/arch/loongarch64/boot/guest_platform/loongarch64/resources.rs
