+++ virtualization/axvm/src/arch/loongarch64/irq.rs
