+++ virtualization/axvm/src/arch/x86_64/boot/images/linux.rs
