+++ virtualization/axvm/src/arch/x86_64/boot/images/x86/boot_params.rs
