+++ virtualization/axvm/src/arch/x86_64/boot/images/x86/linux_boot.rs
