+++ virtualization/axvm/src/boot/mod.rs
@@ -3,29 +3,11 @@
 #[cfg(any(feature = "fs", feature = "host-fs"))]
 use ax_errno::ax_err_type;
 
-pub mod images;
 mod policy;
-
-#[cfg(any(
-    test,
-    target_arch = "aarch64",
-    target_arch = "loongarch64",
-    target_arch = "riscv64"
-))]
-pub mod fdt;
-#[cfg(target_arch = "loongarch64")]
-pub mod guest_platform;
-
-#[cfg(target_arch = "loongarch64")]
-pub use fdt::handle_fdt_operations;
-#[cfg(target_arch = "loongarch64")]
-pub use fdt::init_guest_boot_resources;
-#[cfg(any(target_arch = "aarch64", target_arch = "riscv64"))]
-pub use fdt::{GuestDtbImage, handle_fdt_operations};
-pub use images::{ImageLoader, get_image_header};
-#[cfg(target_arch = "x86_64")]
-pub use images::{is_x86_linux_image_config, x86_qemu_passthrough_block_intx};
+pub use images::*;
 pub use policy::{GuestAcpiTables, GuestBootDescription, GuestDeviceTree, GuestFdtBuilder};
+
+pub use crate::arch::{fdt, guest_platform, images};
 
 /// Build-time image bytes supplied by the hypervisor application.
 #[derive(Clone, Copy, Debug)]
@@ -44,7 +26,6 @@ pub struct StaticVmImage {
 pub trait BootImageProvider {
     fn static_vm_images(&self) -> &'static [StaticVmImage];
 
-    #[cfg(target_arch = "loongarch64")]
     fn static_firmware_images(&self) -> &'static [StaticVmImage] {
         &[]
     }
