+++ virtualization/axvm/src/host/traits.rs
@@ -2,11 +2,6 @@
 
 extern crate alloc;
 
-#[cfg(any(
-    target_arch = "x86_64",
-    target_arch = "aarch64",
-    target_arch = "loongarch64"
-))]
 use alloc::boxed::Box;
 use core::time::Duration;
 
@@ -44,22 +39,15 @@ pub trait HostTime {
     type CancelToken: Copy + Send + Sync + 'static;
 
     /// Convert nanoseconds to hardware ticks.
-    #[cfg(target_arch = "x86_64")]
     fn nanos_to_ticks(&self, nanos: u64) -> u64;
 
     /// Read monotonic host time.
     fn monotonic_time(&self) -> Duration;
 
     /// Program the host one-shot timer.
-    #[cfg(not(target_arch = "loongarch64"))]
     fn set_oneshot_timer(&self, deadline_ns: u64);
 
     /// Register a VM timer callback.
-    #[cfg(any(
-        target_arch = "x86_64",
-        target_arch = "aarch64",
-        target_arch = "loongarch64"
-    ))]
     fn register_timer(
         &self,
         deadline_ns: u64,
@@ -67,7 +55,6 @@ pub trait HostTime {
     ) -> Self::CancelToken;
 
     /// Cancel a VM timer callback.
-    #[cfg(any(target_arch = "x86_64", target_arch = "loongarch64"))]
     fn cancel_timer(&self, token: Self::CancelToken);
 }
 
@@ -84,7 +71,6 @@ pub trait HostCpu {
 }
 
 /// Host console operations.
-#[cfg(target_arch = "x86_64")]
 pub trait HostConsole {
     /// Write raw bytes to host console.
     fn write_bytes(&self, bytes: &[u8]);
