+++ virtualization/axvm/src/manager.rs
@@ -54,7 +54,6 @@ pub fn get_vm_list() -> Vec<AxVMRef> {
 }
 
 /// Run an operation with a VM selected from the process-wide runtime registry.
-#[cfg(any(target_arch = "x86_64", target_arch = "riscv64"))]
 pub(crate) fn with_vm<F, R>(vm_id: VMId, f: F) -> Option<R>
 where
     F: FnOnce(&AxVMRef) -> R,
@@ -64,7 +63,6 @@ where
 }
 
 /// Return the active-vCPU mask for a VM.
-#[cfg(target_arch = "x86_64")]
 pub(crate) fn active_vcpu_mask(vm_id: VMId) -> Option<usize> {
     with_vm(vm_id, |vm| {
         let vcpu_num = vm.vcpu_num();
@@ -77,13 +75,15 @@ pub(crate) fn active_vcpu_mask(vm_id: VMId) -> Option<usize> {
 }
 
 /// Inject a virtual interrupt into a VM's vCPU.
-#[cfg(target_arch = "x86_64")]
 pub(crate) fn inject_interrupt(vm_id: VMId, vcpu_id: usize, vector: usize) -> AxResult {
     crate::runtime::vcpus::queue_interrupt(vm_id, vcpu_id, vector)
 }
 
 /// Inject a virtual interrupt into a VM's vCPU.
-#[cfg(target_arch = "loongarch64")]
+#[expect(
+    dead_code,
+    reason = "only the LoongArch IRQ backend injects external VM interrupts"
+)]
 pub(crate) fn inject_vm_vcpu_interrupt(vm_id: VMId, vcpu_id: usize, vector: usize) -> AxResult {
     use crate::AsVCpuTask;
