+++ virtualization/axvm/src/timer.rs
@@ -3,18 +3,10 @@
 extern crate alloc;
 
 use alloc::boxed::Box;
-#[cfg(any(
-    target_arch = "x86_64",
-    target_arch = "aarch64",
-    target_arch = "loongarch64"
-))]
-use core::sync::atomic::{AtomicUsize, Ordering};
-#[cfg(any(
-    target_arch = "x86_64",
-    target_arch = "aarch64",
-    target_arch = "loongarch64"
-))]
-use core::time::Duration;
+use core::{
+    sync::atomic::{AtomicUsize, Ordering},
+    time::Duration,
+};
 
 use ax_kspin::SpinNoIrq;
 use ax_lazyinit::LazyInit;
@@ -22,33 +14,19 @@ use ax_timer_list::{TimeValue, TimerEvent, TimerList};
 
 use crate::host::{HostTime, default_host};
 
-#[cfg(any(
-    target_arch = "x86_64",
-    target_arch = "aarch64",
-    target_arch = "loongarch64"
-))]
 static TOKEN: AtomicUsize = AtomicUsize::new(0);
 
 struct VmTimerEvent {
-    #[cfg(any(target_arch = "x86_64", target_arch = "loongarch64"))]
     token: usize,
     callback: Box<dyn FnOnce(TimeValue) + Send + 'static>,
 }
 
 impl VmTimerEvent {
-    #[cfg(any(
-        target_arch = "x86_64",
-        target_arch = "aarch64",
-        target_arch = "loongarch64"
-    ))]
     fn new<F>(token: usize, callback: F) -> Self
     where
         F: FnOnce(TimeValue) + Send + 'static,
     {
-        #[cfg(not(any(target_arch = "x86_64", target_arch = "loongarch64")))]
-        let _ = token;
         Self {
-            #[cfg(any(target_arch = "x86_64", target_arch = "loongarch64"))]
             token,
             callback: Box::new(callback),
         }
@@ -64,11 +42,6 @@ impl TimerEvent for VmTimerEvent {
 #[ax_percpu::def_percpu]
 static TIMER_LIST: LazyInit<SpinNoIrq<TimerList<VmTimerEvent>>> = LazyInit::new();
 
-#[cfg(any(
-    target_arch = "x86_64",
-    target_arch = "aarch64",
-    target_arch = "loongarch64"
-))]
 pub(crate) fn register_timer(
     deadline_ns: u64,
     callback: Box<dyn FnOnce(Duration) + Send + 'static>,
@@ -89,7 +62,6 @@ pub(crate) fn register_timer(
     token
 }
 
-#[cfg(any(target_arch = "x86_64", target_arch = "loongarch64"))]
 pub(crate) fn cancel_timer(token: usize) {
     let next_deadline = {
         // SAFETY: The timer list is initialized for each CPU before VM timer
@@ -120,9 +92,6 @@ pub(crate) fn check_events() {
 }
 
 fn rearm_host_timer(next_deadline: Option<TimeValue>) {
-    #[cfg(target_arch = "loongarch64")]
-    let _ = next_deadline;
-    #[cfg(not(target_arch = "loongarch64"))]
     if let Some(deadline) = next_deadline {
         default_host().set_oneshot_timer(deadline.as_nanos() as u64);
     }
@@ -134,6 +103,5 @@ pub(crate) fn init_percpu() {
     // CPU can register VM timers.
     let timer_list = unsafe { TIMER_LIST.current_ref_mut_raw() };
     timer_list.init_once(SpinNoIrq::new(TimerList::new()));
-    #[cfg(target_arch = "loongarch64")]
-    ax_std::os::arceos::modules::ax_task::register_timer_callback(|_| check_events());
+    crate::arch::register_timer_callback();
 }
